c: refactoring -> imports with path alias
