+++ client/src/components/MiniPlayer/index.tsx
@@ -1,12 +1,12 @@
 import React from 'react'
 import styled, { css } from 'styled-components'
-import IconButton from '../../shared/IconButton'
-import ImagePlaceholder from '../../shared/ImagePlaceholder'
-import SvgIcon from '../../shared/SvgIcon'
-import PlayIcon from '../../assets/svg/play.svg?react'
-import PauseIcon from '../../assets/svg/pause.svg?react'
+import SvgIcon from '@/shared/SvgIcon'
+import IconButton from '@/shared/IconButton'
+import PlayIcon from '@/assets/svg/play.svg?react'
+import PauseIcon from '@/assets/svg/pause.svg?react'
+import ImagePlaceholder from '@/shared/ImagePlaceholder'
 import { useReactiveVar } from '@apollo/client'
-import { currentTrackVar } from '../../apollo/cache/variables'
+import { currentTrackVar } from '@/apollo/cache/variables'
 
 interface MiniPlayerProps {
   onPlayerOpen(): void
