+++ client/src/components/Nav/index.tsx
@@ -1,11 +1,11 @@
 import React from 'react'
-import SvgIcon from '../../shared/SvgIcon'
-import { routeNames } from '../../utils/history'
-import SearchIcon from '../../assets/svg/search.svg?react'
-import MusicListIcon from '../../assets/svg/list-music.svg?react'
-import CogIcon from '../../assets/svg/cog.svg?react'
-import { Container, UploadTrackLink, BaseLink } from './Nav.styled'
+import SvgIcon from '@/shared/SvgIcon'
+import SearchIcon from '@/assets/svg/search.svg?react'
+import MusicListIcon from '@/assets/svg/list-music.svg?react'
+import CogIcon from '@/assets/svg/cog.svg?react'
+import { routeNames } from '@/utils/history'
 import { useTranslation } from 'react-i18next'
+import { Container, UploadTrackLink, BaseLink } from './Nav.styled'
 
 const Nav: React.FC = () => {
   const { t } = useTranslation()
