+++ client/src/components/Player/Controls/index.tsx
@@ -1,10 +1,10 @@
 import React from 'react'
-import PlayIcon from '../../../assets/svg/play.svg?react'
-import PauseIcon from '../../../assets/svg/pause.svg?react'
-import ArrowAltIcon from '../../../assets/svg/arrow-alt.svg?react'
-import RepeatIcon from '../../../assets/svg/repeat.svg?react'
-import RandomIcon from '../../../assets/svg/random.svg?react'
-import SvgIcon from '../../../shared/SvgIcon'
+import PlayIcon from '@/assets/svg/play.svg?react'
+import PauseIcon from '@/assets/svg/pause.svg?react'
+import ArrowAltIcon from '@/assets/svg/arrow-alt.svg?react'
+import RepeatIcon from '@/assets/svg/repeat.svg?react'
+import RandomIcon from '@/assets/svg/random.svg?react'
+import SvgIcon from '@/shared/SvgIcon'
 import {
   Container,
   PlayButtonBox,
