+++ client/src/components/Player/TopControls/TopControls.styled.ts
@@ -1,5 +1,5 @@
 import styled, { css } from 'styled-components'
-import IconButton from '../../../shared/IconButton'
+import IconButton from '@/shared/IconButton'
 
 export const Container = styled.div`
   display: flex;
