+++ client/src/pages/Home/NowPlaying/index.tsx
@@ -1,17 +1,18 @@
 import React, { useEffect, useState } from 'react'
-import SvgIcon from '../../../shared/SvgIcon'
-import ImagePlaceholder from '../../../shared/ImagePlaceholder'
-import PlusIcon from '../../../assets/svg/plus.svg?react'
-import TrashIcon from '../../../assets/svg/trash.svg?react'
-import ShareIcon from '../../../assets/svg/share.svg?react'
-import CopyIcon from '../../../assets/svg/copy.svg?react'
-import CheckIcon from '../../../assets/svg/check-circle.svg?react'
+import SvgIcon from '@/shared/SvgIcon'
+import PlusIcon from '@/assets/svg/plus.svg?react'
+import TrashIcon from '@/assets/svg/trash.svg?react'
+import ShareIcon from '@/assets/svg/share.svg?react'
+import CopyIcon from '@/assets/svg/copy.svg?react'
+import CheckIcon from '@/assets/svg/check-circle.svg?react'
+import ImagePlaceholder from '@/shared/ImagePlaceholder'
 import { useReactiveVar } from '@apollo/client'
-import { currentTrackVar, userVar } from '../../../apollo/cache/variables'
+import { currentTrackVar, userVar } from '@/apollo/cache/variables'
 import { useTranslation } from 'react-i18next'
-import { UserMutations, UiMutations } from '../../../apollo/cache/mutations'
-import { useAttachSoundtrackMutation } from '../../../graphql/user/_gen_/attachSoundtrack.mutauion'
-import { useUnattachSoundtrackMutation } from '../../../graphql/user/_gen_/unattachSoundtrack.mutation'
+import { UserMutations, UiMutations } from '@/apollo/cache/mutations'
+import { useAttachSoundtrackMutation } from '@/graphql/user/_gen_/attachSoundtrack.mutauion'
+import { useUnattachSoundtrackMutation } from '@/graphql/user/_gen_/unattachSoundtrack.mutation'
+
 import {
   InnerContainer,
   ImageWrapper,
