+++ client/src/pages/Home/Plugs.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import styled from 'styled-components'
 import { useTranslation } from 'react-i18next'
-import thinkEmojiImage from '../../assets/rastr/thinking.png'
+import thinkEmojiImage from '@/assets/rastr/thinking.png'
 
 const PlugWrapper = styled.div`
   flex-grow: 1;
