+++ client/src/pages/Home/index.tsx
@@ -1,15 +1,15 @@
 import React, { useState, useEffect, useRef, useCallback } from 'react'
-import NowPlaying from './NowPlaying/NowPlaying'
-import PlaylistItem from '../../components/PlaylistItem/PlaylistItem'
-import ScaleLoader from '../../shared/Loader'
-import { userVar, currentTrackVar, userPlaylistVar } from '../../apollo/cache/variables'
-import { useReactiveVar } from '@apollo/client'
-import { UserMutations } from '../../apollo/cache/mutations'
-import { Container, List, MountLoader } from './Home.styled'
+import NowPlaying from './NowPlaying'
+import ScaleLoader from '@/shared/Loader'
+import PlaylistItem from '@/components/PlaylistItem'
 import { NoDataPlug, ErrorPlug } from './Plugs'
-import { Playlist } from '../../apollo/cache/types'
-import { useUserSoundtracksLazyQuery } from '../../graphql/user/_gen_/userSoundtracks.query'
-import type { Soundtrack } from '../../apollo/cache/types'
+import { useReactiveVar } from '@apollo/client'
+import { UserMutations } from '@/apollo/cache/mutations'
+import { userVar, currentTrackVar, userPlaylistVar } from '@/apollo/cache/variables'
+import { useUserSoundtracksLazyQuery } from '@/graphql/user/_gen_/userSoundtracks.query'
+import { Container, List, MountLoader } from './Home.styled'
+import { Playlist } from '@/apollo/cache/types'
+import type { Soundtrack } from '@/apollo/cache/types'
 
 const ITEMS_PER_PAGE = 15
