+++ client/src/pages/UI/Form/Form.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import styled from 'styled-components'
-import TextInput from '../../../shared/FormFields/TextInput'
-import Checkbox from '../../../shared/FormFields/Checkbox'
+import TextInput from '@/shared/FormFields/TextInput'
+import Checkbox from '@/shared/FormFields/Checkbox'
 
 const Container = styled.section``
