+++ client/src/pages/Upload/Audio/Dropzone/Dropzone.styled.ts
@@ -1,5 +1,5 @@
 import styled from 'styled-components'
-import SvgIcon from '../../../../shared/SvgIcon'
+import SvgIcon from '@/shared/SvgIcon'
 
 interface containerStyles {
   isError: boolean
