+++ client/src/pages/Upload/index.tsx
@@ -1,15 +1,15 @@
 import React, { useState, useMemo } from 'react'
 import styled from 'styled-components'
-import Info from './Info/Info'
-import Cover from './Cover/Cover'
-import Audio from './Audio/Audio'
-import { Formik, Form } from 'formik'
-import { useWindowRatio } from '../../hooks'
-import { createTrackStepsSchema } from '../../utils/validationSchemas'
-import { useCreateSoundtrackMutation } from '../../graphql/soundtrack/_gen_/createSoundtrack.mutation'
-import Feedback from './modals/Feedback'
 import Alert from './modals/Alert'
-import history, { routeNames } from '../../utils/history'
+import Feedback from './modals/Feedback'
+import Info from './Info'
+import Cover from './Cover'
+import Audio from './Audio'
+import history, { routeNames } from '@/utils/history'
+import { Formik, Form } from 'formik'
+import { useWindowRatio } from '@/hooks'
+import { createTrackStepsSchema } from '@/utils/validationSchemas'
+import { useCreateSoundtrackMutation } from '@/graphql/soundtrack/_gen_/createSoundtrack.mutation'
 
 enum Step {
   INFO = 0,
