+++ client/src/pages/Upload/modals/Feedback.tsx
@@ -1,11 +1,11 @@
 import React from 'react'
 import styled from 'styled-components'
 import Modal from 'styled-react-modal'
-import sunglassesEmoji from '../../../assets/rastr/tada.png'
-import thinkingEmoji from '../../../assets/rastr/thinking.png'
-import Button from '../../../shared/Button'
+import sunglassesEmoji from '@/assets/rastr/tada.png'
+import thinkingEmoji from '@/assets/rastr/thinking.png'
+import Button from '@/shared/Button'
 import { useTranslation } from 'react-i18next'
-import { ITheme } from '../../../utils/theme'
+import { ITheme } from '@/utils/theme'
 
 interface FeedbackProps {
   isOpen: boolean
