+++ client/src/shared/ImagePlaceholder/index.tsx
@@ -1,8 +1,8 @@
 import React, { useState, useEffect } from 'react'
 import styled, { css } from 'styled-components'
 import LazyLoad from 'react-lazy-load'
-import SvgIcon from '../SvgIcon'
-import PlaceholderImage from '../../assets/svg/music.svg?react'
+import SvgIcon from '@/shared/SvgIcon'
+import PlaceholderImage from '@/assets/svg/music.svg?react'
 import './styles.css'
 
 export interface ImagePlaceholderProps {
