+++ client/src/shared/PrivateRoute/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
-import { routeNames } from '../../utils/history'
+import { routeNames } from '@/utils/history'
 import { useReactiveVar } from '@apollo/client'
-import { isAuthenticatedVar } from '../../apollo/cache/variables'
+import { isAuthenticatedVar } from '@/apollo/cache/variables'
 import { Navigate, Outlet } from 'react-router-dom'
 
 const PrivateRoute: React.FC = () => {
